feat: implement cell filling with a mouse click
