+++ nonograms/scripts/script.js
@@ -1,6 +1,7 @@
-import { createSidebar, currentTemplate, changeTemplate } from "./sidebar.js";
+import { createSidebar } from "./sidebar.js";
 import { createGrid } from "./game-grid.js";
 import { generateRowHints, generateColumnHints } from "./hints.js";
+import { changeTemplate, fillCells } from "./grid-events.js";
 
 export const main = document.createElement("main");
 main.className = "main";
@@ -14,3 +15,4 @@ generateColumnHints();
 createGrid();
 
 changeTemplate();
+fillCells();
